requirements.txt
# API and Data Processing
fastapi==0.104.1
uvicorn==0.23.2
requests==2.31.0
pandas==2.1.1
numpy==1.26.0
yfinance==0.2.31
alpha_vantage==2.3.1
beautifulsoup4==4.12.2
selenium==4.12.0
httpx==0.25.0

# Vector Database and Embeddings
# Using scikit-learn as a simpler alternative to faiss-cpu
scikit-learn==1.6.0
sentence-transformers==2.2.2

# LLM and Agents
together==0.2.5
langchain==0.0.351
crewai==0.1.7
langgraph==0.0.17

# Voice Processing
openai-whisper==20231117
pyttsx3==2.90
SpeechRecognition==3.10.0

# Web UI
streamlit==1.28.0

# Testing and Development
pytest==7.4.3
black==23.10.1
isort==5.12.0

# Deployment
python-dotenv==1.0.0
docker==6.1.3
